new method names .from_unified, .render, and .scan
When working on the NodeJS and Erlang versions of this library it
became apparent these names made more sense.  So, lets make them the
default here as well, and alias the old ones so existing code doesn’t
break.

Update tests to new names and add tests to ensure old method names
continue working as well.